# swagger.json
{
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "swagger": "2.0",
  "info": {
    "description": "Kube API service is service for connection with Kubernetes.",
    "title": "Kube API"
  },
  "paths": {
    "/configmaps": {
      "get": {
        "tags": [
          "ConfigMap"
        ],
        "summary": "Get config maps from all user namespaces.",
        "operationId": "GetSelectedConfigMaps",
        "parameters": [
          {
            "$ref": "#/parameters/UserIDHeader"
          },
          {
            "$ref": "#/parameters/UserRoleHeader"
          },
          {
            "$ref": "#/parameters/UserNamespaceHeader"
          }
        ],
        "responses": {
          "200": {
            "description": "config maps list from all users namespaces",
            "schema": {
              "$ref": "#/definitions/SelectedConfigMapsList"
            }
          },
          "default": {
            "$ref": "#/responses/error"
          }
        },
        "x-method-visibility": "public"
      }
    },
...
    "/namespaces/{namespace}": {
      "get": {
        "tags": [
          "Namespace"
        ],
        "summary": "Get namespace.",
        "operationId": "GetNamespace",
        "parameters": [
          {
            "$ref": "#/parameters/UserIDHeader"
          },
          {
            "$ref": "#/parameters/UserRoleHeader"
          },
          {
            "$ref": "#/parameters/UserNamespaceHeader"
          },
          {
            "type": "string",
            "name": "namespace",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "namespace",
            "schema": {
              "$ref": "#/definitions/NamespaceWithOwner"
            }
          },
          "default": {
            "$ref": "#/responses/error"
          }
        },
        "x-method-visibility": "public"
      },
      "put": {
        "tags": [
          "Namespace"
        ],
        "summary": "Update namespace.",
        "operationId": "UpdateNamespace",
        "parameters": [
          {
            "$ref": "#/parameters/UserIDHeader"
          },
          {
            "$ref": "#/parameters/UserRoleHeader"
          },
          {
            "$ref": "#/parameters/UserNamespaceHeader"
          },
          {
            "type": "string",
            "name": "namespace",
            "in": "path",
            "required": true
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/NamespaceWithOwner"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "namespace updated",
            "schema": {
              "$ref": "#/definitions/NamespaceWithOwner"
            }
          },
          "default": {
            "$ref": "#/responses/error"
          }
        },
        "x-method-visibility": "private"
      },
      "delete": {
        "tags": [
          "Namespace"
        ],
        "summary": "Delete namespace.",
        "operationId": "DeleteNamespace",
        "parameters": [
          {
            "$ref": "#/parameters/UserIDHeader"
          },
          {
            "$ref": "#/parameters/UserRoleHeader"
          },
          {
            "$ref": "#/parameters/UserNamespaceHeader"
          },
          {
            "type": "string",
            "name": "namespace",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "202": {
            "description": "namespace deleted"
          },
          "default": {
            "$ref": "#/responses/error"
          }
        },
        "x-method-visibility": "private"
      }
    },
    "/namespaces/{namespace}/configmaps": {
      "get": {
        "tags": [
          "ConfigMap"
        ],
        "summary": "Get config maps list.",
        "operationId": "GetConfigMapList",
        "parameters": [
          {
            "$ref": "#/parameters/UserIDHeader"
          },
          {
            "$ref": "#/parameters/UserRoleHeader"
          },
          {
            "$ref": "#/parameters/UserNamespaceHeader"
          },
          {
            "type": "string",
            "name": "namespace",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "config maps list",
            "schema": {
              "$ref": "#/definitions/ConfigMapsList"
            }
          },
          "default": {
            "$ref": "#/responses/error"
          }
        },
        "x-method-visibility": "public"
      },
      "post": {
        "tags": [
          "ConfigMap"
        ],
        "summary": "Create config map.",
        "operationId": "CreateConfigMap",
        "parameters": [
          {
            "$ref": "#/parameters/UserIDHeader"
          },
          {
            "$ref": "#/parameters/UserRoleHeader"
          },
          {
            "$ref": "#/parameters/UserNamespaceHeader"
          },
          {
            "type": "string",
            "name": "namespace",
            "in": "path",
            "required": true
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/ConfigMapWithOwner"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "config map created",
            "schema": {
              "$ref": "#/definitions/ConfigMapWithOwner"
            }
          },
          "default": {
            "$ref": "#/responses/error"
          }
        },
        "x-method-visibility": "public"
      }
    },
    "/namespaces/{namespace}/configmaps/{configmap}": {
      "get": {
        "tags": [
          "ConfigMap"
        ],
        "summary": "Get config map.",
        "operationId": "GetConfigMap",
        "parameters": [
          {
            "$ref": "#/parameters/UserIDHeader"
          },
          {
            "$ref": "#/parameters/UserRoleHeader"
          },
          {
            "$ref": "#/parameters/UserNamespaceHeader"
          },
          {
            "type": "string",
            "name": "namespace",
            "in": "path",
            "required": true
          },
          {
            "type": "string",
            "name": "configmap",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "config map",
            "schema": {
              "$ref": "#/definitions/ConfigMapWithOwner"
            }
          },
          "default": {
            "$ref": "#/responses/error"
          }
        },
        "x-method-visibility": "public"
      },
      "put": {
        "tags": [
          "ConfigMap"
        ],
        "summary": "Update config map.",
        "operationId": "UpdateConfigMap",
        "parameters": [
          {
            "$ref": "#/parameters/UserIDHeader"
          },
          {
            "$ref": "#/parameters/UserRoleHeader"
          },
          {
            "$ref": "#/parameters/UserNamespaceHeader"
          },
          {
            "type": "string",
            "name": "namespace",
            "in": "path",
            "required": true
          },
          {
            "type": "string",
            "name": "configmap",
            "in": "path",
            "required": true
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/ConfigMapWithOwner"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "config map updated",
            "schema": {
              "$ref": "#/definitions/ConfigMapWithOwner"
            }
          },
          "default": {
            "$ref": "#/responses/error"
          }
        },
        "x-method-visibility": "public"
      },
      "delete": {
        "tags": [
          "ConfigMap"
        ],
        "summary": "Delete config map.",
        "operationId": "DeleteConfigMap",
        "parameters": [
          {
            "$ref": "#/parameters/UserIDHeader"
          },
          {
            "$ref": "#/parameters/UserRoleHeader"
          },
          {
            "$ref": "#/parameters/UserNamespaceHeader"
          },
          {
            "type": "string",
            "name": "namespace",
            "in": "path",
            "required": true
          },
          {
            "type": "string",
            "name": "configmap",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "202": {
            "description": "config map deleted"
          },
          "default": {
            "$ref": "#/responses/error"
          }
        },
        "x-method-visibility": "public"
      }
    },
...
  },
  "definitions": {
...
    "ConfigMap": {
      "description": "ConfigMap -- model for config map",
      "type": "object",
      "required": [
        "name",
        "data"
      ],
      "properties": {
        "created_at": {
          "description": "creation date in RFC3339 format",
          "type": "string",
          "x-go-name": "CreatedAt"
        },
        "data": {
          "$ref": "#/definitions/ConfigMapData"
        },
        "name": {
          "type": "string",
          "x-go-name": "Name"
        },
        "owner": {
          "type": "string",
          "x-go-name": "Owner"
        }
      },
      "x-go-package": "git.containerum.net/ch/kube-api/vendor/github.com/containerum/kube-client/pkg/model"
    },
    "ConfigMapData": {
      "description": "ConfigMapData -- model for config map data",
      "type": "object",
      "additionalProperties": {
        "type": "string"
      },
      "x-go-package": "git.containerum.net/ch/kube-api/vendor/github.com/containerum/kube-client/pkg/model"
    },
    "ConfigMapsList": {
      "description": "ConfigMapsList -- model for config maps list",
      "type": "object",
      "properties": {
        "configmaps": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ConfigMap"
          },
          "x-go-name": "ConfigMaps"
        }
      },
      "x-go-package": "git.containerum.net/ch/kube-api/vendor/github.com/containerum/kube-client/pkg/model"
    },
...
    "Err": {
      "description": "Err -- standard serializable API error\nMessage -- constant error message:\n+ \"invalid username\"\n+ \"quota exceeded\"\n+ \"validation error\"\n...etc...\nID -- unique error identification code\nDetails -- optional context error messages kinda\n+ \"field 'Replicas' must be non-zero value\"\n+ \"not enough tights to feed gopher\"\n+ \"resource 'God' does't exist\"\nFields -- optional set of key-value pairs",
      "type": "object",
      "properties": {
        "details": {
          "type": "array",
          "items": {
            "type": "string"
          },
          "x-go-name": "Details"
        },
        "fields": {
          "$ref": "#/definitions/Fields"
        },
        "id": {
          "$ref": "#/definitions/ErrID"
        },
        "message": {
          "type": "string",
          "x-go-name": "Message"
        },
        "status_http": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "StatusHTTP"
        }
      },
      "x-go-package": "git.containerum.net/ch/kube-api/vendor/github.com/containerum/cherry"
    },
    "ErrID": {
      "description": "ErrID -- represents unique error ID",
      "type": "object",
      "properties": {
        "kind": {
          "$ref": "#/definitions/ErrKind"
        },
        "sid": {
          "$ref": "#/definitions/ErrSID"
        }
      },
      "x-go-package": "git.containerum.net/ch/kube-api/vendor/github.com/containerum/cherry"
    },
    "ErrKind": {
      "description": "ErrKind -- represents kind of error",
      "type": "integer",
      "format": "uint64",
      "x-go-package": "git.containerum.net/ch/kube-api/vendor/github.com/containerum/cherry"
    },
    "ErrSID": {
      "description": "ErrSID -- represents service ID of error",
      "type": "string",
      "x-go-package": "git.containerum.net/ch/kube-api/vendor/github.com/containerum/cherry"
    },
    "Fields": {
      "description": "Fields -- represents additional key-value fields of error",
      "type": "object",
      "additionalProperties": {
        "type": "string"
      },
      "x-go-package": "git.containerum.net/ch/kube-api/vendor/github.com/containerum/cherry"
    },
...
    "SelectedConfigMapsList": {
      "description": "SelectedConfigMapsList -- model for config maps list from all namespaces",
      "type": "object",
      "additionalProperties": {
        "$ref": "#/definitions/ConfigMapsList"
      },
      "x-go-package": "git.containerum.net/ch/kube-api/vendor/github.com/containerum/kube-client/pkg/model"
    },
...
  },
  "parameters": {
...
    "UserIDHeader": {
      "$ref": "vendor/github.com/containerum/utils/httputil/swagger.json#/parameters/UserIDHeader"
    },
    "UserNamespaceHeader": {
      "$ref": "vendor/github.com/containerum/utils/httputil/swagger.json#/parameters/UserNamespacesHeader"
    },
    "UserRoleHeader": {
      "$ref": "vendor/github.com/containerum/utils/httputil/swagger.json#/parameters/UserRoleHeader"
    }
  },
  "responses": {
    "error": {
      "description": "cherry error",
      "schema": {
        "$ref": "vendor/github.com/containerum/cherry/swagger.json#/definitions/Err"
      }
    }
  }
}